+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # 1.2.3 - In progress
 
 - [REFACTOR] Migrate widgets to WorkManager
-- [Feature] Add mfkey32 attack in debug/options
+- [Feature] Add mfkey32 attack in debug/options + button back
 - [Feature] Add HUB tab
 - [Feature] Wrap more errors with widget
 - [Feature] Add error screen on mfkey32